requirements.txt
watson-developer-cloud
openai
ibm-watson
flask
python-dotenv

